requirements.txt
altair==5.1.1
annotated-types==0.7.0
anyio==4.7.0
asttokens==2.4.0
attrs==23.1.0
backcall==0.2.0
blinker==1.6.2
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
colorama==0.4.6
comm==0.1.4
contourpy==1.3.1
cplex==22.1.1.1
cvxpy==1.3.2
cycler==0.12.1
debugpy==1.8.0
decorator==5.1.1
dlib @ file:///C:/Users/Nikshith/Downloads/Dlib_Windows_Python3.x-main/Dlib_Windows_Python3.x-main/dlib-19.24.1-cp311-cp311-win_amd64.whl#sha256=6f1a5ee167975d7952b28e0ce4495f1d9a77644761cf5720fb66d7c6188ae496
docplex==2.25.236
ecos==2.0.12
et-xmlfile==1.1.0
executing==1.2.0
face-recognition==1.3.0
face_recognition_models==0.3.0
fastapi==0.115.6
filelock==3.16.1
fonttools==4.55.3
fsspec==2024.12.0
gitdb==4.0.10
GitPython==3.1.36
h11==0.14.0
idna==3.4
importlib-metadata==6.8.0
imutils==0.5.4
ipykernel==6.25.2
ipython==8.15.0
jedi==0.19.0
Jinja2==3.1.2
joblib==1.3.2
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
jupyter_client==8.3.1
jupyter_core==5.3.1
kiwisolver==1.4.8
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.10.0
matplotlib-inline==0.1.6
mdurl==0.1.2
mpmath==1.3.0
nest-asyncio==1.5.7
networkx==3.4.2
numpy==1.25.2
opencv-python==4.8.1.78
openpyxl==3.1.2
osqp==0.6.3
packaging==23.1
pandas==2.0.3
parso==0.8.3
pickleshare==0.7.5
Pillow==9.5.0
platformdirs==3.10.0
prompt-toolkit==3.0.39
protobuf==4.24.3
psutil==5.9.5
pure-eval==0.2.2
py-cpuinfo==9.0.0
pyarrow==13.0.0
pydantic==2.10.4
pydantic_core==2.27.2
pydeck==0.8.0
Pygments==2.16.1
Pympler==1.0.1
pyparsing==3.2.1
python-dateutil==2.8.2
python-multipart==0.0.20
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
pywin32==306
PyYAML==6.0.2
pyzmq==25.1.1
qdldl==0.1.7.post0
referencing==0.30.2
requests==2.31.0
rich==13.5.2
rpds-py==0.10.3
scikit-learn==1.3.0
scipy==1.11.1
scs==3.2.3
seaborn==0.13.2
six==1.16.0
smmap==5.0.0
sniffio==1.3.1
stack-data==0.6.2
starlette==0.41.3
sympy==1.13.1
tenacity==8.2.3
threadpoolctl==3.2.0
toml==0.10.2
toolz==0.12.0
torch==2.5.1
torchvision==0.20.1
tornado==6.3.3
tqdm==4.67.1
traitlets==5.9.0
typing_extensions==4.12.2
tzdata==2023.3
tzlocal==4.3.1
ultralytics==8.3.58
ultralytics-thop==2.0.13
urllib3==2.0.4
uvicorn==0.34.0
validators==0.22.0
watchdog==3.0.0
wcwidth==0.2.6
zipp==3.16.2
